% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cleaned_data}
\alias{cleaned_data}
\title{Transport Emissions Data}
\format{
A data frame with 4 variables:
\describe{
\item{entity}{The type of transport, e.g., Bus, Car, etc.}
\item{code}{An optional code column (if applicable)}
\item{year}{The year the data was collected}
\item{emissions_per_km}{Emissions per kilometer, in gCO2/km}
}
}
\source{
Compiled data for Assignment 4
}
\usage{
cleaned_data
}
\description{
A dataset containing emissions per kilometer for different transport modes.
}
\keyword{datasets}
