% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cleaned_data}
\alias{cleaned_data}
\title{Transport Emissions Data}
\format{
A data frame with 4 variables:
\describe{
\item{entity}{Transport mode (e.g., Bus, Diesel Car, etc.)}
\item{code}{An optional code column}
\item{year}{The year the data was collected}
\item{emissions_per_km}{Emissions per kilometer, measured in gCO2/km}
}
}
\source{
Compiled from Assignment Data
}
\usage{
cleaned_data
}
\description{
This dataset contains information about carbon emissions per kilometer for various transport modes.
}
\keyword{datasets}
